requirements.txt
beautifulsoup4==4.8.2
requests==2.25.1
selenium==3.141.0
mysql-connector==2.2.9